services/zrun-bff/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zrun-bff"
version = "0.1.0-dev"
description = "BFF service for zrun (OAuth2 + JWT re-issuance)"
requires-python = ">=3.14"
dependencies = [
    "zrun-core",
    "zrun-schema",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-jose[cryptography]",
    "python-multipart",
    "itsdangerous",
    "grpcio",
]

[project.scripts]
zrun-bff = "zrun_bff.main:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-benchmark>=4.0.0",
    "httpx",
]

[tool.uv.sources]
zrun-core = { workspace = true }
zrun-schema = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/zrun_bff"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    # Ignore DeprecationWarning from python-jose library (uses datetime.utcnow())
    "ignore::DeprecationWarning:jose",
]
